Fixed issue from previous commit, CL 56459004.
BUG=4550
R=stefan@webrtc.org

Review URL: https://codereview.webrtc.org/1168783002

Cr-Commit-Position: refs/heads/master@{#9393}

+++ webrtc/modules/remote_bitrate_estimator/test/estimators/nada_unittest.cc
@@ -557,16 +557,16 @@ TEST_F(NadaReceiverSideTest, PacketLossWideGap) {
 
 // Packets arriving unordered should not be counted as losted.
 TEST_F(NadaReceiverSideTest, PacketLossUnorderedPackets) {
-  int num_packets = nada_receiver_.GetSetCapacity() / 2;
+  size_t num_packets = nada_receiver_.GetSetCapacity() / 2;
   std::vector<uint16_t> sequence_numbers;
 
-  for (int i = 0; i < num_packets; ++i) {
+  for (size_t i = 0; i < num_packets; ++i) {
     sequence_numbers.push_back(static_cast<uint16_t>(i + 1));
   }
 
   random_shuffle(sequence_numbers.begin(), sequence_numbers.end());
 
-  for (int i = 0; i < num_packets; ++i) {
+  for (size_t i = 0; i < num_packets; ++i) {
     const MediaPacket media_packet(kFlowId, 0, 0, sequence_numbers[i]);
     // Arrival time = 0, all packets will be considered.
     nada_receiver_.ReceivePacket(0, media_packet);
